+++ src/components/language.vue
@@ -6,8 +6,8 @@
       title="Change language"
       >
       <q-icon
-        size="md"
         :name="fasLanguage"
+        size="md"
       />
 
       <q-menu
@@ -15,14 +15,13 @@
         fit
         self="center middle"
         >
-        <!-- FIXME: remove inline style -->
-        <q-list style="min-width: 100px;">
+        <q-list>
           <q-item clickable v-close-popup>
-            <q-item-section>EN</q-item-section>
+            <q-item-section>DE</q-item-section>
           </q-item>
 
           <q-item clickable v-close-popup>
-            <q-item-section>DE</q-item-section>
+            <q-item-section>{{ lang }}</q-item-section>
           </q-item>
         </q-list>
       </q-menu>
@@ -35,6 +34,11 @@ import { fasLanguage } from '@quasar/extras/fontawesome-v5';
 
 export default {
   name: 'Language',
+  data() {
+    return {
+      lang: 'EN',
+    };
+  },
   created() {
     this.fasLanguage = fasLanguage;
   },