+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2020-09-29
+
+### Added
+
+- Add config for unit tests via jest. testfiles goto *tests/unit/specs/*test.js*
+
+### Changed
+
+- delete tab-folder with components. provided a computed prop to wait for conditional changes
+
 ## [1.4.0] - 2020-09-11
 
 ### Added